refactor: implement handle validation by zod
