+++ product-service/lambda-functions/createProduct.ts
@@ -2,7 +2,8 @@ import { APIGatewayProxyEvent } from "aws-lambda";
 import { DynamoDBDocument } from "@aws-sdk/lib-dynamodb";
 import { DynamoDB } from "@aws-sdk/client-dynamodb";
 import { randomUUID } from "crypto";
-import { sendResponse } from "./sendResponse";
+import { sendResponse } from "./helpers/sendResponse";
+import { TProductDto, validateProductDto } from "./helpers/validateProductDto";
 
 const PRODUCTS_TABLE_NAME = process.env.PRODUCTS_TABLE_NAME || "";
 const STOCK_TABLE_NAME = process.env.STOCK_TABLE_NAME || "";
@@ -18,18 +19,26 @@ exports.handler = async (event: APIGatewayProxyEvent) => {
       message: "Invalid request, you are missing the parameter body",
     });
 
-  const item =
+  const productDto =
     typeof event.body == "object" ? event.body : JSON.parse(event.body);
 
+  const isValidDto = validateProductDto(productDto);
+  if (!isValidDto)
+    return sendResponse(400, {
+      message: "Invalid request body",
+    });
+
   const id = randomUUID();
+
+  const { title, description, price, count } = productDto as TProductDto;
 
   const productsParams = {
     TableName: PRODUCTS_TABLE_NAME,
     Item: {
       id,
-      title: item.title,
-      description: item.description,
-      price: item.price,
+      title,
+      description,
+      price,
     },
   };
 
@@ -37,7 +46,7 @@ exports.handler = async (event: APIGatewayProxyEvent) => {
     TableName: STOCK_TABLE_NAME,
     Item: {
       productId: id,
-      count: item.count,
+      count,
     },
   };
 
@@ -48,7 +57,8 @@ exports.handler = async (event: APIGatewayProxyEvent) => {
   try {
     await db.transactWrite(transactParams);
     return sendResponse(201, {
-      message: "Product is created",
+      ...productDto,
+      id,
     });
   } catch (dbError) {
     return sendResponse(500, { message: `DynamoDB Error: ${dbError}` });
