+++ product-service/lambda-functions/helpers/sendResponse.ts
