+++ product-service/lambda-functions/helpers/validateProductDto.ts
@@ -1,0 +1,16 @@
+import { z } from "zod";
+
+const ProductSchema = z.object({
+  title: z.string(),
+  description: z.string(),
+  price: z.number().min(0),
+  count: z.number().min(0),
+}).strict();
+
+export type TProductDto = z.infer<typeof ProductSchema>;
+
+export const validateProductDto = (productDto: unknown) => {
+    const validationResult = ProductSchema.safeParse(productDto);
+
+    return validationResult.success;
+};
